[wip] refactor: 各種処理を適切なファイルに分割
