--- app/models/api_payload.py
@@ -1,21 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class ActivateSubscriptionAPIPayload:
-    guild_id: int
-
-
-@dataclass
-class RenewSubscriptionAPIPayload:
-    new_plan: str
-
-
-@dataclass
-class CheckoutSessionAPIPayload:
-    plan: str
-
-
-@dataclass
-class PutDictAPIPayload:
-    dict: dict
