+++ app/routes/webhooks/stripe_webhook.py
@@ -5,7 +5,8 @@ import asyncpg.connection
 import stripe
 from fastapi import APIRouter, Request, Response, HTTPException
 
-from app.constants import STRIPE_WEBHOOK_SECRET, MONTHLY1_PRICE_ID, YEARLY1_PRICE_ID, MONTHLY2_PRICE_ID, YEARLY2_PRICE_ID
+from app.constants import STRIPE_WEBHOOK_SECRET, MONTHLY1_PRICE_ID, YEARLY1_PRICE_ID, MONTHLY2_PRICE_ID, \
+    YEARLY2_PRICE_ID
 from app.db.connection import get_connection_pool
 
 router = APIRouter()
