+++ app/schemas/api_response.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from app.models.character_usage import CharacterUsages
+
+
+@dataclass
+class CharacterUsageAPIResponse:
+    message: str
+    data: CharacterUsages
