+++ app/schemas/character_usage.py
@@ -1,0 +1,10 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+from app.models.character_usage import CharacterUsage
+
+
+class CharacterUsagesUpdate(BaseModel):
+    wavenet: Optional[CharacterUsage] = None
+    standard: Optional[CharacterUsage] = None
