+++ app/schemas/checkout_session.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class CheckoutSessionCreate(BaseModel):
+    plan: str
