+++ app/schemas/guild_dict.py
@@ -1,0 +1,7 @@
+from typing import Dict, Any
+
+from pydantic import BaseModel, Field
+
+
+class DictPut(BaseModel):
+    dict: Dict[str, Any] = Field(...)
