+++ app/schemas/subscription.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class SubscriptionActivate(BaseModel):
+    guild_id: int
+
+
+class SubscriptionRenew(BaseModel):
+    new_plan: str
